requirements.txt
Django
Pillow
python-decouple
PyYAML
requests
djangorestframework
djangorestframework-simplejwt